requirements.txt
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.5.0; python_version >= '3.7'
anyio==3.7.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
dataclasses-json==0.6.1; python_version >= '3.7' and python_version < '4.0'
fastapi==0.103.2
frozenlist==1.4.0; python_version >= '3.8'
greenlet==3.0.0; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
langchain==0.0.308
langsmith==0.0.42; python_version < '4.0' and python_full_version >= '3.8.1'
marshmallow==3.20.1; python_version >= '3.8'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.0; python_version < '3.13' and python_version >= '3.9'
openai==0.28.1
packaging==23.2; python_version >= '3.7'
pillow==10.0.1
pydantic==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.21; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
tenacity==8.2.3; python_version >= '3.7'
tiktoken==0.5.1
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==2.0.6; python_version >= '3.7'
uvicorn==0.23.2
yarl==1.9.2; python_version >= '3.7'
